fix(Sales Order): ensure that items are pre-filled
Due to a recent change in frappe, the Year of Settlement will be pre-filled in a new form, when it was previously selected as list view filter. In this case, we need to trigger a field change event for Year of Settlement and ensure that the item query was set beforehand.